Change service id to uint16, add init checks

+++ type_definitions/primitives.hpp
@@ -7,7 +7,7 @@
 #define DEFAULT_SERVICE_ID 0
 #define DEFAULT_SECRET CryptoPP::SecByteBlock()
 
-typedef int service_id_t;
+typedef uint16_t service_id_t;
 typedef uint16_t member_id_t;
 typedef uint8_t message_id_t;
 typedef CryptoPP::SecByteBlock blinded_secret_t;
